Split unicode handling logic into separate plugin
